requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
transparent-background>=1.3.4
numpy>=1.24.0
opencv-python>=4.8.0
torch>=1.13.0
torchvision>=0.14.0
python-dotenv>=1.0.0
aiofiles>=23.2.1
uvloop>=0.19.0
requests>=2.31.0
